Improved Replies security

+++ src/GamingWiki.Web/Controllers/RepliesController.cs
@@ -1,6 +1,7 @@
 ﻿using GamingWiki.Services.Contracts;
 using GamingWiki.Web.Infrastructure;
 using GamingWiki.Web.Models.Replies;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GamingWiki.Web.Controllers
@@ -13,6 +14,7 @@ namespace GamingWiki.Web.Controllers
         => this.helper = helper;
 
         [HttpPost]
+        [Authorize]
         public IActionResult Add(ReplyFormModel model, int commentId)
         {
             if (!this.ModelState.IsValid)
@@ -26,8 +28,21 @@ namespace GamingWiki.Web.Controllers
            return this.Redirect($"/Articles/Details?articleId={articleId}");
         }
 
+        [Authorize]
         public IActionResult Delete(int replyId)
         {
+            if (!this.helper.ReplyExists(replyId))
+            {
+                return this.View("Error");
+            }
+
+            var replierId = this.helper.GetReplyAuthorId(replyId);
+
+            if (!this.User.IsAdmin() && this.User.GetId() != replierId)
+            {
+                return this.Unauthorized();
+            }
+
             var articleId = this.helper.Delete(replyId);
             return this.Redirect($"/Articles/Details?articleId={articleId}");
         }
